holy > death finally works > state change vfx > hot/cold visual indicators > sfx needs tweaking > critcial state/drill power > hitstop on explosions > immunity visual >score pop up needs fixing > discharge to capsule (needs adjusting) > juice and fine tuning
all in all I'm happy with progress
need a lot more experience for full vision

+++ Assets/Scripts/LEGACY SCRIPT/BulletController.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class BulletController : MonoBehaviour
 {
-    public float Damage = 1f;
+    public int Damage = 1;
     public GameObject explosionFX;
